# idl/grpcbasic0.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "grpcbasic0.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/user": {
      "post": {
        "operationId": "RecordUser",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/idlUser"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/idlUserRecordReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/v1/users": {
      "post": {
        "operationId": "ListUsers",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/idlUsers"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/idlUsersListReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/v1/users/{id}": {
      "post": {
        "operationId": "GetUser",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/idlUser"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/idlUserGetReq"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    }
  },
  "definitions": {
    "idlUser": {
      "type": "object",
      "properties": {
        "age": {
          "type": "string",
          "format": "int64"
        },
        "fortune": {
          "type": "string",
          "format": "string"
        },
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "idlUserGetReq": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "idlUserRecordReq": {
      "type": "object",
      "properties": {
        "age": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "idlUsers": {
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/idlUser"
          }
        }
      }
    },
    "idlUsersListReq": {
      "type": "object",
      "properties": {
        "count": {
          "type": "string",
          "format": "int64"
        },
        "down": {
          "type": "boolean",
          "format": "boolean"
        },
        "start": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  }
}